Updated the project to most up-to-date version. Shifted dungeon scenes into the SceneFactory.java class.
